[#666] package SettingsView
