+++ ios/CCC/UI/Settings/SettingsItemView.swift
@@ -1,0 +1,62 @@
+//
+//  SettingsItemView.swift
+//  CCC
+//
+//  Created by Mustafa Ozhan on 23.04.22.
+//  Copyright © 2022 orgName. All rights reserved.
+//
+
+import SwiftUI
+import Resources
+
+struct SettingsItemView: View {
+
+    let imgName: String
+    let title: String
+    let subTitle: String
+    let value: String
+    let onClick: () -> Void
+
+    var body: some View {
+        HStack {
+            Image(systemName: imgName)
+                .frame(width: 48, height: 48, alignment: .center)
+                .font(.system(size: 24))
+                .imageScale(.large)
+                .accentColor(MR.colors().text.get())
+                .padding(.bottom, 8)
+                .padding(.top, 8)
+
+            VStack {
+                HStack {
+                    Text(title).font(.title3)
+                    Spacer()
+                }
+
+                Spacer()
+
+                HStack {
+                    Text(subTitle).font(.footnote)
+                    Spacer()
+                }
+            }.frame(height: 30)
+
+            Spacer()
+
+            Text(value)
+                .lineLimit(2)
+                .multilineTextAlignment(.trailing)
+                .font(.caption)
+
+            Image(systemName: "chevron.right")
+                .frame(width: 48, height: 48, alignment: .center)
+                .imageScale(.large)
+                .accentColor(MR.colors().text.get())
+        }
+        .listRowInsets(.init())
+        .listRowBackground(MR.colors().background.get())
+        .contentShape(Rectangle())
+        .onTapGesture { onClick() }
+        .lineLimit(1)
+    }
+}
