+++ ios/CCC/UI/Settings/SettingsView.swift
@@ -172,70 +172,3 @@ struct SettingsView: View {
         }
     }
 }
-
-struct SettingsToolbarView: View {
-    var backEvent: () -> Void
-
-    var body: some View {
-        HStack {
-            ToolbarButton(clickEvent: backEvent, imgName: "chevron.left")
-
-            Text(MR.strings().txt_settings.get())
-                .font(.title3)
-
-            Spacer()
-        }.padding(EdgeInsets(top: 20, leading: 10, bottom: 5, trailing: 20))
-    }
-}
-
-struct SettingsItemView: View {
-
-    let imgName: String
-    let title: String
-    let subTitle: String
-    let value: String
-    let onClick: () -> Void
-
-    var body: some View {
-        HStack {
-            Image(systemName: imgName)
-                .frame(width: 48, height: 48, alignment: .center)
-                .font(.system(size: 24))
-                .imageScale(.large)
-                .accentColor(MR.colors().text.get())
-                .padding(.bottom, 8)
-                .padding(.top, 8)
-
-            VStack {
-                HStack {
-                    Text(title).font(.title3)
-                    Spacer()
-                }
-
-                Spacer()
-
-                HStack {
-                    Text(subTitle).font(.footnote)
-                    Spacer()
-                }
-            }.frame(height: 30)
-
-            Spacer()
-
-            Text(value)
-                .lineLimit(2)
-                .multilineTextAlignment(.trailing)
-                .font(.caption)
-
-            Image(systemName: "chevron.right")
-                .frame(width: 48, height: 48, alignment: .center)
-                .imageScale(.large)
-                .accentColor(MR.colors().text.get())
-        }
-        .listRowInsets(.init())
-        .listRowBackground(MR.colors().background.get())
-        .contentShape(Rectangle())
-        .onTapGesture { onClick() }
-        .lineLimit(1)
-    }
-}
